pyproject.toml
[tool.poetry]
name = "ytvideosummarizer"
version = "0.1.0"
description = "YouTube Video Summarizer"
authors = ["karuna-sth <karuna.sth01@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.1"
drf-yasg = "^1.21.7"
djangorestframework = "^3.15.2"
assemblyai = "^0.34.0"
python-dotenv = "^1.0.1"
openai = "^1.53.0"
psycopg2-binary = "^2.9.10"
pytube3 = "^9.6.4"
pytubefix = "^8.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
